Error handling

+++ app/javascript/src/parts/logout_btn.vue
@@ -4,15 +4,26 @@
 
 <script>
 import axios from 'axios'
+import { mapMutations } from 'vuex'
 
 export default {
   name: 'LogOutBtn',
 
   methods: {
+    ...mapMutations([
+      'setErrors',
+    ]),
+
     async logout () {
-      await axios.delete('/api/v1/users/sign_out')
-      window.sessionStorage.clear()
-      window.location.pathname = '/'
+      let [response, errors] = await this.handle(axios.delete('/api/v1/users/sign_out'))
+      if (errors) {
+        this.setErrors(errors)
+      }
+      else {
+        window.sessionStorage.clear()
+        window.localStorage.clear()
+        window.location.pathname = '/'
+      }
     }
   }
 }
